// src/app/model/ListQuestion.ts
export interface ListQuestion {
  Package_ID: string;
  Question: string;
  Question_ID: string;
  Type?: string;
  Type_ID: string;
  DateCreate: string;
  Answer_Text: string;
  Answer: string;
  Package_Name: string;
  OptionA: string;
  OptionB: string;
  OptionC: string;
  OptionD: string;
  OptionE: string;
  Note: string;
  Image: string;
  Category: string;
  Explanation: string;
}